add background to #content div
it's gradient doesn't go all the way down either. We must keep
the backgrounds on #wait, #delay, and #error, since they are
needed to cover the #formWrap background.

the same background of #error is added to #content, so that
even though #error doesn't extend all the way down, the gradient
still does, and it looks correct on Android again.

fixes 1721